+++ CHANGELOG.md
@@ -27,9 +27,19 @@ November 7th, 2022
 - Added logic and behavior for planes.
     - Planes occasionally fly across the screen to make the level feel more alive.
     - Planes are randomly generated at the beginning of each game.
-    - Planes currently don't have any practical functionality.
+    - Planes currently don't have any practical functionality, and are disabled in game.
     - Currently the planes simply use water bottle sprites as placeholders.
 - Changed trash spawning behavior.
     - Decreased the trash drop size each round from 40 to 20.
     - Decreased the round length from 10 seconds to 7.
     - Increased the number of rounds from 3 to 6.
+
+
+## Version 1.0
+
+### Factual Update
+
+*Release date to be determined*
+
+- Added accurate statistics to the end screen.
+    - The end screen now displays the number of bottles picked up, approximate number of animals saved, and approximate years worth of plastic kept out of the ocean.